Add ESLint Prettier plugin
This will does 2 things:

- ESLint Prettier Config is just a config by prettier that turns off a
lot of the rules that conflicts with prettier, which is IMO a terrible
choice of a config name.
- Having Prettier as an ESLint config will ensure that there will be no
conflict changes between before and after running the `yarn format`
command, especially when it actually changes a folder that the developer
is not working on.
e.g. if i'm just changing something in src/danhSomeone, I don't want
`yarn format` to affect src/somethingElse.