Remove node_modules volume
This is causing more trouble than its worth. A lot of times people find
it frustrating and confusing when the host folder has the correct
node_modules, but the one in the container isn't, causing some scripts
unable to run.

This change will fix that, by making the code volume sync with the
container.

Pros: The node_modules folder will always be in sync, and
the user may not need to run bash inside the container to execute
commands after the initial setup.

Cons: The user needs to run the `yarn install` command twice while
setting up the repo for the first time. Once on the host machine, and
once while building the container.

But imo the pros outweigh the cons here.