Replace alpine base images with distroless
This will keep the production image size down, while staying on top of
security issues, and debugging debian issues are way easier than dealing
alpine based images, which the distroless base images are using.

This will also remove a lot of the pre-config steps that we need to get
the container going on alpine. So despite gaining a bit in size in the
dev image, but the production one is a lot smaller.