Import mocked from jest-mock instead of ts-jest
This is because the mocked util has been merged into the main jest
library. But the @types/jest package is not updated to have the
mocked function yet, so this will have to be imported from jest-mock.

When the update goes through, this will be updated to be importing
from jest directly.