Remove running development in container
This proves to be more trouble than its worth at the moment, because
essentially we have 2 copies of node_modules that are not always in
sync because of different runtime environments between the container
(Linux) and the host machine (which can be either Linux, Mac or
Windows). The guide will be updated to reflect that.